+++ packages/lambda/src/common/logger.ts
@@ -7,6 +7,7 @@ export class Logger {
   constructor() {
     // WIP minimalistic logger
     this.logger = pino({
+      base: null,
       formatters: { level: (level) => ({ level }) },
       level: process.env.LOG_LEVEL || 'info',
     });
@@ -18,7 +19,16 @@ export class Logger {
     });
   }
 
-  public log(message: string): void {
+  public debug(message: string): void {
+    this.logger.debug(message);
+  }
+  public info(message: string): void {
     this.logger.info(message);
   }
 }
+
+// test out the logger
+const logger = new Logger();
+// @ts-ignore
+logger.debug('debug');
+logger.info('info');
