+++ packages/lambda/src/index.ts
@@ -1,1 +1,1 @@
-export * from './toolkit.js';
+export * from './toolkit/index.js';
