+++ packages/lambda/src/toolkit/toolkit.ts
@@ -6,51 +6,33 @@ import type {
   ICreatable,
   ICreatableOptions,
   Toolkit,
+  ToolkitOptions,
 } from './toolkit.types.js';
+
+import { isConstructable } from '../utils/index.js';
 
 /**
  * Creates a new toolkit to be used throughout your serverless application.
  * @param name the name of the app
- * @returns the tools that can be utilized:
- * ```txt
- *   • name: the name of your app
- *   • logger: a json logger
- *   • cache: in-memory cache
- *   • create: abstract utility factory
- * ```
+ * @param options optional parameters for the toolkit
+ * @returns a toolkit that can be used to create tools
  */
-export function createToolkit(name: string): Toolkit {
-  const logger = {} as Logger;
-  const cache = {} as Cache;
+export function createToolkit(name: string, options?: ToolkitOptions): Toolkit {
+  const logger = (options?.logger || {}) as Logger;
+  const cache = (options?.cache || {}) as Cache;
 
   // eslint-disable-next-line jsdoc/require-jsdoc
   function create<I extends ICreatable, O extends ICreatableOptions>(
     tool: Creatable<I, O>,
     options?: O
   ): I {
-    if (isConstructable(tool)) {
-      return new tool(logger, cache, create, options);
-    }
+    if (isConstructable(tool)) return new tool(logger, cache, create, options);
 
     return tool(logger, cache, create, options);
   }
 
   return {
-    cache,
     create,
-    logger,
     name,
   };
-}
-
-/**
- * Attempts to determine if a passed function is one that can be invoked with
- * the `new` keyword.
- * @param fn The function to test.
- * @returns The resulting boolean test.
- */
-function isConstructable(
-  fn?: Creatable<ICreatable, any>
-): fn is { new (...args: any[]): any } {
-  return !!fn && !!fn.prototype && !!fn.prototype.constructor;
 }
