+++ packages/lambda/src/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './function.js';
+export * from './object.js';
